Only drop table if it exists

+++ db/migrate/20201021174413_drop_stocks.rb
@@ -1,6 +1,6 @@
 class DropStocks < ActiveRecord::Migration[5.1]
   def up
-    drop_table :stocks
+    drop_table :stocks if table_exists? :stocks
   end
 
   def down
